requirements.txt
python-box==6.0.2
coloredlogs==15.0.1
pandas
numpy
seaborn
ensure
PyYAML
pydantic
plotly
scikit-learn
catboost
xgboost
-e .